// src/app/shared/interfaces/kyc.ts
export interface Kyc {
  first_name: string;
  email:string;
  middle_name: string;
  last_name: string;
  gender: any;
  dob: any;
  phonecode: any;
  phone: any;
  address: any;
  city: any;
  state: any;
  country: any;
  zipcode: any;
  doc_type: any;
  doc_number: any;
  doc_exp_date: any;
  p_res_type: any;
  // father_husband_name: string;
  // relation: string;
}